+++ CMS.Service/Exceptions/BadRequestException.cs
@@ -7,12 +7,12 @@ namespace CMS.Service.Exceptions
     {
         public BadRequestException() : base()
         {
-            Error = new BaseResult { StatusCode = (int)HttpStatusCode.BadRequest, Message = "Geçersiz istek." };
+            Error = new BaseResult { StatusCode = HttpStatusCode.BadRequest, Message = "Geçersiz istek." };
         }
 
         public BadRequestException(string message) : base(message)
         {
-            Error = new BaseResult { StatusCode = (int)HttpStatusCode.BadRequest, Message = message };
+            Error = new BaseResult { StatusCode = HttpStatusCode.BadRequest, Message = message };
         }
 
         protected override HttpStatusCode HttpStatusCode => HttpStatusCode.BadRequest;
