+++ CMS.Service/Exceptions/FoundException.cs
@@ -10,12 +10,12 @@ namespace CMS.Service.Exceptions
     {
         public FoundException() : base()
         {
-            Error = new BaseResult { StatusCode = (int)HttpStatusCode.Found, Message = "Kayıt Mevcut." };
+            Error = new BaseResult { StatusCode = HttpStatusCode.Found, Message = "Kayıt Mevcut." };
         }
 
         public FoundException(string message) : base(message)
         {
-            Error = new BaseResult { StatusCode = (int)HttpStatusCode.Found, Message = message };
+            Error = new BaseResult { StatusCode = HttpStatusCode.Found, Message = message };
         }
 
         protected override HttpStatusCode HttpStatusCode => HttpStatusCode.Found;
