+++ CMS.Service/Exceptions/NotAcceptableException.cs
@@ -10,12 +10,12 @@ namespace CMS.Service.Exceptions
     {
         public NotAcceptableException() : base()
         {
-            Error = new BaseResult { StatusCode = (int)HttpStatusCode.NotAcceptable, Message = "Kabul Edilmeyen İstek." };
+            Error = new BaseResult { StatusCode = HttpStatusCode.NotAcceptable, Message = "Kabul Edilmeyen İstek." };
         }
 
         public NotAcceptableException(string message) : base(message)
         {
-            Error = new BaseResult { StatusCode = (int)HttpStatusCode.NotAcceptable, Message = message };
+            Error = new BaseResult { StatusCode = HttpStatusCode.NotAcceptable, Message = message };
         }
 
         protected override HttpStatusCode HttpStatusCode => HttpStatusCode.NotAcceptable;
